feat: Send Swagger docs (#2445)
* chore: Move send to module

* chore: Consolidate existing docs

* chore: Move to subfolders

* feat: Document and add validation layer to create send events endpoint

* test: Update test import refs

* feat: Donwload application files

* fix: Linting after rebase
