+++ api.planx.uk/modules/pay/service/inviteToPay/createPaymentSendEvents.ts
@@ -1,11 +1,11 @@
 import { ComponentType } from "@opensystemslab/planx-core/types";
 import { NextFunction, Request, Response } from "express";
 import { gql } from "graphql-request";
-import { $api, $public } from "../../../../client";
 import {
-  ScheduledEventResponse,
+  CombinedResponse,
   createScheduledEvent,
 } from "../../../../lib/hasura/metadata";
+import { $api, $public } from "../../../../client";
 import { getMostRecentPublishedFlow } from "../../../../helpers";
 import { Flow, Node, Team } from "../../../../types";
 
@@ -13,12 +13,6 @@ enum Destination {
   BOPS = "bops",
   Uniform = "uniform",
   Email = "email",
-}
-
-interface CombinedResponse {
-  bops?: ScheduledEventResponse;
-  uniform?: ScheduledEventResponse;
-  email?: ScheduledEventResponse;
 }
 
 // Create "One-off Scheduled Events" in Hasura when a payment request is paid
