+++ api.planx.uk/modules/send/bops/bops.test.ts
@@ -1,10 +1,10 @@
 import nock from "nock";
 import supertest from "supertest";
-import { queryMock } from "../tests/graphqlQueryMock";
-import app from "../server";
-import { expectedPayload } from "../tests/mocks/bopsMocks";
+import { queryMock } from "../../../tests/graphqlQueryMock";
+import app from "../../../server";
+import { expectedPayload } from "../../../tests/mocks/bopsMocks";
 
-jest.mock("../modules/saveAndReturn/service/utils", () => ({
+jest.mock("../../saveAndReturn/service/utils", () => ({
   markSessionAsSubmitted: jest.fn(),
 }));
