+++ api.planx.uk/modules/send/utils/helpers.ts
@@ -1,6 +1,6 @@
 import { gql } from "graphql-request";
-import airbrake from "../airbrake";
-import { $api } from "../client";
+import airbrake from "../../../airbrake";
+import { $api } from "../../../client";
 
 export async function logPaymentStatus({
   sessionId,
